Move a few modules and functions around to minimize unnecessary imports of CLI routines.
