+++ splitgraph/ingestion/singer/commandline/BUILD
@@ -1,0 +1,1 @@
+python_library()
