+++ splitgraph/ingestion/singer/commandline/__init__.py
